backend/requirements.txt
fastapi==0.104.0
uvicorn==0.23.2
pydantic==2.4.2
python-dotenv==1.0.0
python-multipart==0.0.6
pydantic-settings==2.0.3